+++ CHANGELOG.md
@@ -16,6 +16,12 @@ Semantic versioning in our case means:
   But, in the future we might change the configuration names/logic,
   change the client facing API, change code conventions significantly, etc.
 
+## WIP
+
+### Bugfixes
+
+- Fixes the false positive `WPS222` for nested conditions, #3630
+
 
 ## 1.6.1